[LongCat-Image] Address PR #5170 review: bf16 + skip
- Switch all three component tests from fp32 to bf16, matching the
  flux/flux2/wan component-test convention and halving DRAM footprint.
  VAE decoder verified passing on n150 in bf16 (95.86s).
- Convert transformer + text_encoder from xfail to skip so CI does not
  waste compute on a known single-chip OOM (tracking #5169, #5168).
  Both confirmed still OOM on single n150 at bf16; skip reasons carry
  the observed bf16 allocation figures.

+++ tests/torch/models/longcat_image/test_vae_decoder.py
@@ -18,8 +18,8 @@ def test_vae_decoder():
     torch.manual_seed(42)
 
     loader = ModelLoader(ModelVariant.VAE)
-    model = loader.load_model(dtype_override=torch.float32)
-    inputs = loader.load_inputs(dtype_override=torch.float32)
+    model = loader.load_model(dtype_override=torch.bfloat16)
+    inputs = loader.load_inputs(dtype_override=torch.bfloat16)
 
     run_graph_test(
         model,
